+++ packages/peer-store/src/utils/to-peer-pb.ts
@@ -2,7 +2,7 @@
 import { publicKeyToProtobuf } from '@libp2p/crypto/keys'
 import { InvalidParametersError } from '@libp2p/interface'
 import { equals as uint8ArrayEquals } from 'uint8arrays/equals'
-import { dedupeFilterAndSortAddresses } from './dedupe-addresses.js'
+import { dedupeFilterAndSortAddresses } from './dedupe-addresses.ts'
 import type { AddressFilter } from '../index.js'
 import type { Tag, Peer as PeerPB } from '../pb/peer.js'
 import type { ExistingPeer } from '../store.js'