+++ packages/crypto/CHANGELOG.md
@@ -17,6 +17,43 @@
   * dependencies
     * @libp2p/interface bumped from ^1.0.2 to ^1.1.0
 
+## [5.1.17](https://github.com/libp2p/js-libp2p/compare/crypto-v5.1.16...crypto-v5.1.17) (2026-04-16)
+
+
+### Bug Fixes
+
+* support Node.js type stripping by importing from .ts files ([#3448](https://github.com/libp2p/js-libp2p/issues/3448)) ([8623584](https://github.com/libp2p/js-libp2p/commit/86235842b7aaae7fe1a5fb33f19e1fe5e9c34e3b))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^3.2.1 to ^3.2.2
+
+## [5.1.16](https://github.com/libp2p/js-libp2p/compare/crypto-v5.1.15...crypto-v5.1.16) (2026-04-14)
+
+
+### Bug Fixes
+
+* support require ([#3449](https://github.com/libp2p/js-libp2p/issues/3449)) ([f21efd1](https://github.com/libp2p/js-libp2p/commit/f21efd121a9d3755f28049541c1b0758e3498683))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^3.2.0 to ^3.2.1
+
+## [5.1.15](https://github.com/libp2p/js-libp2p/compare/crypto-v5.1.14...crypto-v5.1.15) (2026-04-08)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^3.1.1 to ^3.2.0
+
 ## [5.1.14](https://github.com/libp2p/js-libp2p/compare/crypto-v5.1.13...crypto-v5.1.14) (2026-03-27)