+++ packages/metrics-prometheus/CHANGELOG.md
@@ -81,6 +81,52 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [5.0.18](https://github.com/libp2p/js-libp2p/compare/prometheus-metrics-v5.0.17...prometheus-metrics-v5.0.18) (2026-04-16)
+
+
+### Bug Fixes
+
+* support Node.js type stripping by importing from .ts files ([#3448](https://github.com/libp2p/js-libp2p/issues/3448)) ([8623584](https://github.com/libp2p/js-libp2p/commit/86235842b7aaae7fe1a5fb33f19e1fe5e9c34e3b))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^3.2.1 to ^3.2.2
+  * devDependencies
+    * @libp2p/logger bumped from ^6.2.5 to ^6.2.6
+    * @libp2p/utils bumped from ^7.0.16 to ^7.0.17
+
+## [5.0.17](https://github.com/libp2p/js-libp2p/compare/prometheus-metrics-v5.0.16...prometheus-metrics-v5.0.17) (2026-04-14)
+
+
+### Bug Fixes
+
+* support require ([#3449](https://github.com/libp2p/js-libp2p/issues/3449)) ([f21efd1](https://github.com/libp2p/js-libp2p/commit/f21efd121a9d3755f28049541c1b0758e3498683))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^3.2.0 to ^3.2.1
+  * devDependencies
+    * @libp2p/logger bumped from ^6.2.4 to ^6.2.5
+    * @libp2p/utils bumped from ^7.0.15 to ^7.0.16
+
+## [5.0.16](https://github.com/libp2p/js-libp2p/compare/prometheus-metrics-v5.0.15...prometheus-metrics-v5.0.16) (2026-04-08)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^3.1.1 to ^3.2.0
+  * devDependencies
+    * @libp2p/logger bumped from ^6.2.3 to ^6.2.4
+    * @libp2p/utils bumped from ^7.0.14 to ^7.0.15
+
 ## [5.0.15](https://github.com/libp2p/js-libp2p/compare/prometheus-metrics-v5.0.14...prometheus-metrics-v5.0.15) (2026-03-27)