+++ packages/protocol-autonat-v2/CHANGELOG.md
@@ -1,5 +1,59 @@
 # Changelog
 
+## [2.0.17](https://github.com/libp2p/js-libp2p/compare/autonat-v2-v2.0.16...autonat-v2-v2.0.17) (2026-04-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^3.2.1 to ^3.2.2
+    * @libp2p/interface-internal bumped from ^3.1.1 to ^3.1.2
+    * @libp2p/peer-collections bumped from ^7.0.16 to ^7.0.17
+    * @libp2p/utils bumped from ^7.0.16 to ^7.0.17
+  * devDependencies
+    * @libp2p/crypto bumped from ^5.1.16 to ^5.1.17
+    * @libp2p/logger bumped from ^6.2.5 to ^6.2.6
+    * @libp2p/peer-id bumped from ^6.0.7 to ^6.0.8
+
+## [2.0.16](https://github.com/libp2p/js-libp2p/compare/autonat-v2-v2.0.15...autonat-v2-v2.0.16) (2026-04-14)
+
+
+### Bug Fixes
+
+* support require ([#3449](https://github.com/libp2p/js-libp2p/issues/3449)) ([f21efd1](https://github.com/libp2p/js-libp2p/commit/f21efd121a9d3755f28049541c1b0758e3498683))
+
+
+### Dependencies
+
+* update p-retry to 8.x.x ([#3442](https://github.com/libp2p/js-libp2p/issues/3442)) ([aaf3fbb](https://github.com/libp2p/js-libp2p/commit/aaf3fbbe224bb7c62dfb9a2ffc8751794a93daec))
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^3.2.0 to ^3.2.1
+    * @libp2p/interface-internal bumped from ^3.1.0 to ^3.1.1
+    * @libp2p/peer-collections bumped from ^7.0.15 to ^7.0.16
+    * @libp2p/utils bumped from ^7.0.15 to ^7.0.16
+  * devDependencies
+    * @libp2p/crypto bumped from ^5.1.15 to ^5.1.16
+    * @libp2p/logger bumped from ^6.2.4 to ^6.2.5
+    * @libp2p/peer-id bumped from ^6.0.6 to ^6.0.7
+
+## [2.0.15](https://github.com/libp2p/js-libp2p/compare/autonat-v2-v2.0.14...autonat-v2-v2.0.15) (2026-04-08)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^3.1.1 to ^3.2.0
+    * @libp2p/interface-internal bumped from ^3.0.14 to ^3.1.0
+    * @libp2p/peer-collections bumped from ^7.0.14 to ^7.0.15
+    * @libp2p/utils bumped from ^7.0.14 to ^7.0.15
+  * devDependencies
+    * @libp2p/crypto bumped from ^5.1.14 to ^5.1.15
+    * @libp2p/logger bumped from ^6.2.3 to ^6.2.4
+    * @libp2p/peer-id bumped from ^6.0.5 to ^6.0.6
+
 ## [2.0.14](https://github.com/libp2p/js-libp2p/compare/autonat-v2-v2.0.13...autonat-v2-v2.0.14) (2026-03-27)